requirements.txt
flask
flask_cors
pandas
matplotlib
seaborn
scikit-learn
torch
textblob
gensim
networkx
transformers
huggingface_hub
nltk
pyLDAvis==3.4.1
sqlalchemy
pymysql
mysql-connector-python
spacy==3.5.3
numpy==1.24.2
blis==0.7.9
gunicorn